% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateGeometricFeatures.R
\name{calculateGeometricFeatures}
\alias{calculateGeometricFeatures}
\title{calculateGeometricFeatures}
\usage{
calculateGeometricFeatures(inMatrix, radius = 1, progress = T)
}
\arguments{
\item{inMatrix}{a matrix with columns of X Y Z coordinates,
and rows of observations}

\item{radius}{maximum radius for which to find nearest neighbours}

\item{progress}{a logical boolean - if true a progress bar
will be shown}
}
\value{
matrix with geometric features
}
\description{
calculateGeometricFeatures
}
\examples{
calculateGeometricFeatures(lidar,5, TRUE)
}
