% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateGeometricFeatures.R
\name{calculateGeometricFeatures__}
\alias{calculateGeometricFeatures__}
\title{calculateGeometricFeatures__}
\usage{
calculateGeometricFeatures__(inMatrix, idx, progress = T)
}
\arguments{
\item{inMatrix}{a matrix with columns of X Y Z coordinates,
and rows of observations, so M rows and 3 columns}

\item{idx}{a matrix with nearest neighbour indices of the inMatrix cloud,
so M x N where N is the number of nearest neighbours (k value in RANN library)}

\item{progress}{a logical boolean - if true a progress bar
will be shown}
}
\value{
matrix with geometric features
}
\description{
calculateGeometricFeatures__
}
\examples{
nn <- calculateNeighboursIDX(lidar[1:100,],5, TRUE)
nf <- calculateGeometricFeatures(lidar[1:100,],nn, TRUE)
}
