% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateGeometricFeatures.R
\name{geometricFeaturesCalculate}
\alias{geometricFeaturesCalculate}
\title{Title}
\usage{
geometricFeaturesCalculate(matrix, progress = T)
}
\arguments{
\item{matrix}{a matrix with columns of X Y Z coordinates,
and rows of observations}

\item{progress}{a logical boolean - if true a progress bar
will be shown}
}
\value{
matrix with geometric features
}
\description{
Title
}
\examples{
#no example
}
