% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lidar.R
\docType{data}
\name{lidar}
\alias{lidar}
\title{LiDAR Dataset n.1.}
\format{
A data frame with many rows and 3 columns, X Y and Z:
\describe{
\item{X}{X coordinate}
\item{Y}{Y coordinate}
\item{Z}{height above sea level}
}
}
\source{
\url{https://github.com/LAStools/LAStools/raw/master/data/fusa.laz/}
}
\usage{
lidar
}
\description{
Example Lidar dataset from LASTools (fusa.las) with only 3d data
}
\keyword{datasets}
