% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateGeometricFeatures.R
\name{calcGF}
\alias{calcGF}
\title{calcGF}
\usage{
calcGF(inMatrix, radius = 1, progress = T, threads = 0)
}
\arguments{
\item{inMatrix}{a matrix with columns of X Y Z coordinates,
and rows of observations, so M rows and 3 columns}

\item{radius}{maximum radius for which to find nearest neighbours}

\item{progress}{a logical bolean - if true a progress bar
will be shown}

\item{threads}{number of threads. If zero (default) is set, it will use t-2,
i.e. all threads except 2 if machine has more than 4 threads. Only works if
CloudGeometry was build with OpenMP support (see documentation)}
}
\value{
matrix with M rows and 3 columns with the geometric
features (3 eigen values in ascending order)
}
\description{
Calculate Geometric Features
}
\examples{
#subset the first 100 rows of the lidar point cloud example to limit execution time.
nn <- calcGF(lidar[1:100,],5, TRUE)
}
