% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateGeometricFeatures.R
\name{calculateGeometricFeatures}
\alias{calculateGeometricFeatures}
\title{calculateGeometricFeatures}
\usage{
calculateGeometricFeatures(inMatrix, progress = T)
}
\arguments{
\item{inMatrix}{a matrix with columns of X Y Z coordinates,
and rows of observations}

\item{progress}{a logical boolean - if true a progress bar
will be shown}
}
\value{
matrix with geometric features
}
\description{
calculateGeometricFeatures
}
\examples{
#calculateGeometricFeatures(matrix(1:12, ncol = 3, nrow = 4), TRUE)
}
