% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateGeometricFeatures.R
\name{calculateGeometricFeatures}
\alias{calculateGeometricFeatures}
\title{calculateNeighboursIDX}
\usage{
calculateGeometricFeatures(inMatrix, radius = 1, progress = TRUE)
}
\arguments{
\item{inMatrix}{a matrix with columns of X Y Z coordinates,
and rows of observations, so M rows and 3 columns}

\item{radius}{maximum radius for which to find nearest neighbours}

\item{progress}{a logical boolean - if true a progress bar
will be shown}
}
\value{
integer matrix with id values M x 100 (k=100) which is the
number of nearest neighbours
}
\description{
calculateNeighboursIDX
}
\examples{
#subset the first 100 rows of the lidar point cloud example to limit execution time.
nf <- calculateGeometricFeatures(lidar[1:100,],5, TRUE)
}
