% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lidar.R
\docType{data}
\name{lidar}
\alias{lidar}
\title{LiDAR Dataset.}
\format{
A data frame with many rows and 3 columns, X Y and Z:
\describe{
\item{X}{X coordinate}
\item{Y}{Y coordinate}
\item{Z}{height above sea level}
}
}
\source{
\url{https://lastools.github.io/}
}
\usage{
lidar
}
\description{
Attractiveness judgements of same-sex and other-sex self-resembling faces from DeBruine (2004).
}
\keyword{datasets}
